Add API to read strings without pre-allocating
The new function `json_read_str_part()` works in a similar manner to
`json_read_str()` except it will also return success on partial reads.
Before the first iteration, the caller needs to initialize the `more`
parameter to false.  After the first and subsequent calls, the caller
can qualify a succesful iteration by reading this value to see if the
entire string was read or not.  If more space is required, the caller
can reallocate the string however they choose and pass in the increased
capacity on the following iteration.

To avoid making a lot of strlen() calls, the function also takes an
in/out length parameter.  All the caller needs to do is initialize this
value to 0 before the first call and then leave it alone.