Add methods to write numbers with set precision
Works nice when paired with a rounding method when you want to write no
more than a certain number of digits.

We can't use `%g` for these because it prohibits the values from being
round-tripped.  The meaning of precision we want is "number of places
after the decimal in non-exponent form".  It's unfortunate that this
will cause the serialization size for large values to increase
significantly, but I don't think the time to add special rules avoiding
that in some cases is worthwhile.